requirements.txt
python-dotenv==0.19.0
discord.py==1.7.3
heroku3==4.2.3
numpy==1.21.2
opencv-python-headless==4.5.3.56
python-Levenshtein==0.12.2
fuzzywuzzy==0.18.0
Unidecode==1.2.0
psycopg2-binary==2.9.1
SQLAlchemy==1.4.23
validators==0.18.2